loadouts make do betterer* (*maybe)
